Move get_keys, get_first_key, valid_ip, and in_network to shared/utils.py
